backend/requirements.txt
# Django
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# CORS
django-cors-headers==4.3.1

# Environment
python-dotenv==1.0.0

# Google Gemini AI
google-generativeai==0.3.2

# Async tasks
celery==5.3.4
redis==5.0.1

# Image handling
Pillow==10.1.0

# Utilities
python-dateutil==2.8.2
requests==2.31.0

# Development
django-extensions==3.2.3
